perf: halve fused5 dQ workspace traffic

+++ include/fused_bwd_dq_reduce.h
@@ -10,7 +10,7 @@ namespace shaobo::fa3::bwd::fused_bwd {
 
 size_t dq_workspace_bytes(const ShaoboFa3Params* params);
 
-int launch_dq_reduction(const float* partial,
+int launch_dq_reduction(const __half* partial,
                         __half* dq,
                         const ShaoboFa3Params* params);
